requirements.txt
numpy==1.21.0
torchvision==0.10.0
tqdm==4.61.2
pandas==1.3.0
dagshub==0.1.6
requests==2.26.0
torch==1.9.0
h5py==3.3.0
apex==0.9.10dev
Pillow==8.3.1
PyYAML==5.4.1
scikit_learn==0.24.2
wandb==0.10.33